refactor(renderer): split the workspace view along context boundaries
`WorkspaceShell` was the god controller one layer up: 430 lines calling
eighteen `use*()` hooks so it could assemble every pane's props. Each
pane now reads the domains it renders and nothing else, and the shell
reads only the six that decide which pane renders.

AppView.tsx 855 -> 130 lines; WorkspaceShell 430 -> 87, 18 contexts -> 6.
